+++ tests/cmds/test_users.py
@@ -1,9 +1,11 @@
 import json
 
 import pytest
+from pandas import DataFrame
 from py42.response import Py42Response
 from requests import Response
 
+from code42cli.cmds.users import _add_legal_hold_membership_to_user_dataframe
 from code42cli.main import cli
 
 
@@ -32,6 +34,41 @@ TEST_USERS_RESPONSE = {
         }
     ]
 }
+TEST_MATTER_RESPONSE = {
+    "legalHolds": [
+        {"legalHoldUid": "123456789", "name": "Legal Hold #1", "active": True},
+        {"legalHoldUid": "987654321", "name": "Legal Hold #2", "active": True},
+    ]
+}
+TEST_CUSTODIANS_RESPONSE = {
+    "legalHoldMemberships": [
+        {
+            "legalHoldMembershipUid": "99999",
+            "active": True,
+            "creationDate": "2020-07-16T08:50:23.405Z",
+            "legalHold": {"legalHoldUid": "123456789", "name": "Legal Hold #1"},
+            "user": {
+                "userUid": "911162111513111325",
+                "username": "test.username@example.com",
+                "email": "test.username@example.com",
+                "userExtRef": None,
+            },
+        },
+        {
+            "legalHoldMembershipUid": "11111",
+            "active": True,
+            "creationDate": "2020-07-16T08:50:23.405Z",
+            "legalHold": {"legalHoldUid": "987654321", "name": "Legal Hold #2"},
+            "user": {
+                "userUid": "911162111513111325",
+                "username": "test.username@example.com",
+                "email": "test.username@example.com",
+                "userExtRef": None,
+            },
+        },
+    ]
+}
+
 TEST_EMPTY_USERS_RESPONSE = {"users": []}
 TEST_USERNAME = TEST_USERS_RESPONSE["users"][0]["username"]
 TEST_USER_ID = TEST_USERS_RESPONSE["users"][0]["userId"]
@@ -48,6 +85,14 @@ def _create_py42_response(mocker, text):
 
 def get_all_users_generator():
     yield TEST_USERS_RESPONSE
+
+
+def matter_list_generator():
+    yield TEST_MATTER_RESPONSE
+
+
+def custodian_list_generator():
+    yield TEST_CUSTODIANS_RESPONSE
 
 
 @pytest.fixture
@@ -68,6 +113,18 @@ def get_user_id_success(cli_state):
 @pytest.fixture
 def get_user_id_failure(cli_state):
     cli_state.sdk.users.get_by_username.return_value = TEST_EMPTY_USERS_RESPONSE
+
+
+@pytest.fixture
+def get_all_matter_success(cli_state):
+    cli_state.sdk.legalhold.get_all_matters.return_value = matter_list_generator()
+
+
+@pytest.fixture
+def get_all_custodian_success(cli_state):
+    cli_state.sdk.legalhold.get_all_matter_custodians.return_value = (
+        custodian_list_generator()
+    )
 
 
 @pytest.fixture
@@ -167,6 +224,32 @@ def test_list_users_when_given_excluding_active_and_inactive_uses_active_equals_
     cli_state.sdk.users.get_all.assert_called_once_with(
         active=None, org_uid=None, role_id=None
     )
+
+
+def test_add_legal_hold_membership_to_user_dataframe_adds_legal_hold_columns_to_dataframe(
+    cli_state, get_all_matter_success, get_all_custodian_success
+):
+    testdf = DataFrame.from_records(
+        [{"userUid": "840103986007089121", "status": "Active"}]
+    )
+    result = _add_legal_hold_membership_to_user_dataframe(cli_state.sdk, testdf)
+    assert "legalHoldUid" in result.columns
+    assert "legalHoldName" in result.columns
+
+
+def test_list_include_legal_hold_membership_merges_in_and_concats_legal_hold_info(
+    runner,
+    cli_state,
+    get_all_users_success,
+    get_all_custodian_success,
+    get_all_matter_success,
+):
+    result = runner.invoke(
+        cli, ["users", "list", "--include-legal-hold-membership"], obj=cli_state
+    )
+
+    assert "Legal Hold #1,Legal Hold #2" in result.output
+    assert "123456789,987654321" in result.output
 
 
 def test_add_user_role_adds(